refactored and added integration test for transactions api with in-memory h2 database

+++ src/main/java/com/retailer/rewardsprogramapi/entity/Transaction.java
@@ -1,7 +1,5 @@
 package com.retailer.rewardsprogramapi.entity;
 
-import java.math.BigInteger;
-
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.Data;
@@ -10,8 +8,8 @@ import lombok.Data;
 @Data
 public class Transaction {
 	@Id
-	private BigInteger transactionId;
-	private BigInteger customerId;
+	private Long transactionId;
+	private Long customerId;
 	private Integer transactionDate;
 	private Float transactionAmount;
